Main/GuestSession: build fix (please review).
svn:sync-xref-src-repo-rev: r79908

+++ src/VBox/Main/include/GuestSessionImpl.h
@@ -1,5 +1,5 @@
 
-/* $Id: GuestSessionImpl.h 79899 2012-08-08 22:56:44Z andreas.loeffler@oracle.com $ */
+/* $Id: GuestSessionImpl.h 79908 2012-08-09 06:11:32Z noreply@oracle.com $ */
 /** @file
  * VirtualBox Main - XXX.
  */
@@ -70,7 +70,7 @@ public:
                       const Utf8Str &strSource, const Utf8Str &strDest, uint32_t uFlags);
 
     SessionTaskCopyTo(GuestSession *pSession,
-                      PRTFILE pSourceFile, size_t cbSourceOffset, size_t cbSourceSize,
+                      PRTFILE pSourceFile, size_t cbSourceOffset, uint64_t cbSourceSize,
                       const Utf8Str &strDest, uint32_t uFlags);
 
     virtual ~SessionTaskCopyTo(void);
